Update main.css
Adjusted CSS and markup so that everything displays in a single column.  Updated containers with fixed widths set smaller than 100%. Replaced them with relative widths of 100% so that things stretch to fit across any viewport width.
Adjusted touch targets to make them easy to hit by adding padding of 1.5 em to all the a tags.